# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/v1",
    "paths": {
        "/feiras": {
            "get": {
                "description": "Displays all feiras registered on the API.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Feiras"
                ],
                "summary": "List feiras",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the distrito",
                        "name": "distrito",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Name of regiao5",
                        "name": "regiao5",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Name of feira",
                        "name": "nomefeira",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Name of bairro",
                        "name": "bairro",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.FeiraLivre"
                            }
                        }
                    },
                    "204": {
                        "description": "Feiras not found"
                    }
                }
            },
            "post": {
                "description": "Creates a new feira",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Feiras"
                ],
                "summary": "Create a feira",
                "parameters": [
                    {
                        "description": "Object for persisting the feira",
                        "name": "content",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.FeiraLivrePersist"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.FeiraLivre"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/feiras/registro/{registro}": {
            "delete": {
                "description": "Delete a feira by id",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Feiras"
                ],
                "summary": "Delete a feira",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The identifier for the feira",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    },
                    "204": {
                        "description": "Feira not found"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/feiras/{id}": {
            "get": {
                "description": "Get a feira",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Feiras"
                ],
                "summary": "Get a feira",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The identifier for the feira",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.FeiraLivre"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "description": "Updates a feira",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Feiras"
                ],
                "summary": "Update a feira",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The identifier for the feira",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Object for updating a feira",
                        "name": "content",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.FeiraLivreUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "description": "Delete a feira by id",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Feiras"
                ],
                "summary": "Delete a feira",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The identifier for the feira",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    },
                    "204": {
                        "description": "Feira not found"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.FeiraLivre": {
            "type": "object",
            "properties": {
                "areap": {
                    "type": "integer"
                },
                "bairro": {
                    "type": "string"
                },
                "codDist": {
                    "type": "integer"
                },
                "codSubPref": {
                    "type": "integer"
                },
                "distrito": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lat": {
                    "type": "integer"
                },
                "logradouro": {
                    "type": "string"
                },
                "long": {
                    "type": "integer"
                },
                "nomeFeira": {
                    "type": "string"
                },
                "numero": {
                    "type": "string"
                },
                "referencia": {
                    "type": "string"
                },
                "regiao5": {
                    "type": "string"
                },
                "regiao8": {
                    "type": "string"
                },
                "registro": {
                    "type": "string"
                },
                "setCens": {
                    "type": "integer"
                },
                "subPref": {
                    "type": "string"
                }
            }
        },
        "domain.FeiraLivrePersist": {
            "type": "object",
            "properties": {
                "areap": {
                    "type": "integer"
                },
                "bairro": {
                    "type": "string"
                },
                "codDist": {
                    "type": "integer"
                },
                "codSubPref": {
                    "type": "integer"
                },
                "distrito": {
                    "type": "string"
                },
                "lat": {
                    "type": "integer"
                },
                "logradouro": {
                    "type": "string"
                },
                "long": {
                    "type": "integer"
                },
                "nomeFeira": {
                    "type": "string"
                },
                "numero": {
                    "type": "string"
                },
                "referencia": {
                    "type": "string"
                },
                "regiao5": {
                    "type": "string"
                },
                "regiao8": {
                    "type": "string"
                },
                "registro": {
                    "type": "string"
                },
                "setCens": {
                    "type": "integer"
                },
                "subPref": {
                    "type": "string"
                }
            }
        },
        "domain.FeiraLivreUpdate": {
            "type": "object",
            "properties": {
                "areap": {
                    "type": "integer"
                },
                "bairro": {
                    "type": "string"
                },
                "codDist": {
                    "type": "integer"
                },
                "codSubPref": {
                    "type": "integer"
                },
                "distrito": {
                    "type": "string"
                },
                "lat": {
                    "type": "integer"
                },
                "logradouro": {
                    "type": "string"
                },
                "long": {
                    "type": "integer"
                },
                "nomeFeira": {
                    "type": "string"
                },
                "numero": {
                    "type": "string"
                },
                "referencia": {
                    "type": "string"
                },
                "regiao5": {
                    "type": "string"
                },
                "regiao8": {
                    "type": "string"
                },
                "setCens": {
                    "type": "integer"
                },
                "subPref": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "carrierIdAuthentication": {
            "type": "apiKey",
            "name": "carrierId",
            "in": "header"
        }
    }
}